Moved stack under commands
