+++ pkg/commands/stack/lint.go
@@ -6,7 +6,6 @@ import (
 	"io/ioutil"
 	"strings"
 
-	"github.com/gimlet-io/gimlet-cli/pkg/stack/template"
 	"github.com/urfave/cli/v2"
 	"github.com/xeipuuv/gojsonschema"
 	"gopkg.in/yaml.v3"
@@ -36,17 +35,17 @@ func lint(c *cli.Context) error {
 		return fmt.Errorf("cannot read stack config file: %s", err.Error())
 	}
 
-	var stackConfig template.StackConfig
+	var stackConfig StackConfig
 	err = yaml.Unmarshal(stackConfigYaml, &stackConfig)
 	if err != nil {
 		return fmt.Errorf("cannot parse stack config file: %s", err.Error())
 	}
 
-	stackDefinitionYaml, err := template.StackDefinitionFromRepo(stackConfig.Stack.Repository)
+	stackDefinitionYaml, err := StackDefinitionFromRepo(stackConfig.Stack.Repository)
 	if err != nil {
 		return fmt.Errorf("cannot get stack definition: %s", err.Error())
 	}
-	var stackDefinition template.StackDefinition
+	var stackDefinition StackDefinition
 	err = yaml.Unmarshal([]byte(stackDefinitionYaml), &stackDefinition)
 	if err != nil {
 		return fmt.Errorf("cannot parse stack definition: %s", err.Error())
